+++ libs/ui/src/lib/spinner/__stories__/Spinner.stories.mdx
@@ -1,0 +1,20 @@
+<!-- Spinner.stories.mdx -->
+
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks'
+import { Default, Styled } from './Spinner.stories'
+
+<Meta title="Components/Spinner" />
+
+# Spinner
+
+<Canvas>
+  <Story story={Default} />
+</Canvas>
+
+## Styling
+
+Styling gets passed through, so you can use the `tw` or `css` props to style as desired.
+
+<Canvas>
+  <Story name="Styled with tw=" story={Styled} />
+</Canvas>
