+++ libs/ui/src/lib/spinner/__stories__/Spinner.stories.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import 'twin.macro'
+
+import Spinner from '../Spinner'
+
+export const Default = () => <Spinner />
+
+export const Styled = () => <Spinner tw="text-yellow text-5xl" />
